[DEV-3539] Adds links to Twitter, Facebook, Github, Linkedin on Footer (#1518)
* [DEV-3539] adding the social icons

* [DEV-3539] adding the new styles

* [DEV-3539] adding modal redirect

* [DEV-3539] Moving to containers dir

* [DEV-3539] updating imports

* [DEV-3539] fixing order of social links
