+++ src/_scss/layouts/default/footer/footer.scss
@@ -19,7 +19,7 @@
         @include pad(0 0 rem(30));
         max-width: $site-max-width;
         @include media($footer-horizontal) {
-            @include pad(0 rem(30) rem(60));
+            @include pad(0 rem(30) rem(40));
             @include display(flex);
             @include align-items(flex-start);
             @include justify-content(flex-start);
@@ -27,8 +27,74 @@
 
         @import "./_footerLogo";
         @import "./_footerLinks";
-    }
+        
+        @include flex-wrap(wrap);
 
+        .legal-and-social-links {
+            list-style-type: none;
+            width: 100%;
+            padding: 0;
+            margin-top: rem(73);
+            @include display(flex);
+            @include align-items(center);
+            @include justify-content(center);
+            @include flex-direction(column);
+            @include media($footer-horizontal) {
+                @include justify-content(flex-start);
+                @include flex-direction(row);
+            }
+        
+            .copyright__legal-item {
+                @include flex(0 1 auto);
+                margin-right: rem(10);
+                @include media($footer-horizontal) {
+                    &::before {
+                        content: '|';
+                        font-size: 80%;
+                        color: $color-gray-light;
+                        margin-right: rem(10);
+                    }
+                    &:first-child {
+                        &::before {
+                            content: '';
+                            font-size: 0;
+                            margin-right: 0;
+                        }
+                    }
+                }
+                .copyright__link {
+                    font-size: rem(16);
+                    color:  darken( #fff, 20% );
+                    text-decoration: none;
+                    &:hover, &:active {
+                        color: $color-white;
+                    }
+                }
+            }
+
+            .social-link {
+                &:nth-child(4) {
+                    @include media($footer-horizontal) {
+                        margin-left: auto;
+                    }
+                }
+                &:last-child,
+                & {
+                    margin-bottom: rem(5);
+                    padding: 0 rem(15);
+                }
+                &:hover {
+                    button svg {
+                        color: white;
+                    }
+                }
+                button {
+                    @include button-unstyled;
+                    font-size: rem(32);
+                }
+            }
+        }
+    }
 
     @import "./_footerCopyright";
 }
