+++ src/js/components/about/DBInfo.jsx
@@ -6,10 +6,10 @@
 import React from 'react';
 
 import * as MetaTagHelper from 'helpers/metaTagHelper';
+import Footer from 'containers/Footer';
 
 import MetaTags from '../sharedComponents/metaTags/MetaTags';
 import Header from '../sharedComponents/header/Header';
-import Footer from '../sharedComponents/Footer';
 
 require('pages/dbInfo/dbInfoPage.scss');
