+++ src/js/components/about/legal/common/LegalPage.jsx
@@ -11,11 +11,11 @@ import {
     privacyPageMetaTags,
     foiaPageMetaTags
 } from 'helpers/metaTagHelper';
+import Footer from 'containers/Footer';
 
 import MetaTags from 'components/sharedComponents/metaTags/MetaTags';
 import Header from 'components/sharedComponents/header/Header';
 import StickyHeader from 'components/sharedComponents/stickyHeader/StickyHeader';
-import Footer from 'components/sharedComponents/Footer';
 
 import LegalContent from './LegalContent';
