+++ src/js/components/account/Account.jsx
@@ -7,10 +7,10 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import * as MetaTagHelper from 'helpers/metaTagHelper';
+import Footer from 'containers/Footer';
 
 import MetaTags from '../sharedComponents/metaTags/MetaTags';
 import Header from '../sharedComponents/header/Header';
-import Footer from '../sharedComponents/Footer';
 
 import AccountHeader from './AccountHeader';
 import AccountOverview from './AccountOverview';
