+++ src/js/components/errorPage/ErrorPage.jsx
@@ -8,10 +8,10 @@ import React from 'react';
 import { ExclamationCircle } from 'components/sharedComponents/icons/Icons';
 
 import * as MetaTagHelper from 'helpers/metaTagHelper';
+import Footer from 'containers/Footer';
 
 import MetaTags from '../sharedComponents/metaTags/MetaTags';
 import Header from '../sharedComponents/header/Header';
-import Footer from '../sharedComponents/Footer';
 
 require('pages/errorPage/errorPage.scss');
