+++ src/js/components/explorer/ExplorerWrapperPage.jsx
@@ -8,10 +8,10 @@ import PropTypes from 'prop-types';
 
 import { explorerPageMetaTags } from 'helpers/metaTagHelper';
 
+import Footer from 'containers/Footer';
 import MetaTags from 'components/sharedComponents/metaTags/MetaTags';
 import Header from 'components/sharedComponents/header/Header';
 import StickyHeader from 'components/sharedComponents/stickyHeader/StickyHeader';
-import Footer from 'components/sharedComponents/Footer';
 
 const propTypes = {
     children: PropTypes.element
