+++ src/js/components/keyword/KeywordPage.jsx
@@ -12,6 +12,7 @@ import * as MetaTagHelper from 'helpers/metaTagHelper';
 import * as MoneyFormatter from 'helpers/moneyFormatter';
 
 import ResultsTableContainer from 'containers/keyword/table/ResultsTableContainer';
+import Footer from 'containers/Footer';
 import BulkDownloadModalContainer from
     'containers/bulkDownload/modal/BulkDownloadModalContainer';
 import DownloadButton from 'components/search/header/DownloadButton';
@@ -19,7 +20,6 @@ import DownloadButton from 'components/search/header/DownloadButton';
 import MetaTags from '../sharedComponents/metaTags/MetaTags';
 import Header from '../sharedComponents/header/Header';
 import StickyHeader from '../sharedComponents/stickyHeader/StickyHeader';
-import Footer from '../sharedComponents/Footer';
 
 import KeywordSearchBar from './KeywordSearchBar';
