+++ src/js/components/search/SearchPage.jsx
@@ -11,12 +11,12 @@ import * as MetaTagHelper from 'helpers/metaTagHelper';
 
 import FullDownloadModalContainer from
     'containers/search/modals/fullDownload/FullDownloadModalContainer';
+import Footer from 'containers/Footer';
 import DownloadButton from 'components/search/header/DownloadButton';
 
 import MetaTags from '../sharedComponents/metaTags/MetaTags';
 import Header from '../sharedComponents/header/Header';
 import StickyHeader from '../sharedComponents/stickyHeader/StickyHeader';
-import Footer from '../sharedComponents/Footer';
 import SearchSidebar from './SearchSidebar';
 import SearchResults from './SearchResults';
